BleManager split into 2 classes
