+++ ble/src/main/java/no/nordicsemi/android/ble/ReadRequest.java
@@ -26,6 +26,7 @@ import android.bluetooth.BluetoothDevice;
 import android.bluetooth.BluetoothGatt;
 import android.bluetooth.BluetoothGattCharacteristic;
 import android.bluetooth.BluetoothGattDescriptor;
+import android.os.Handler;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
@@ -68,8 +69,15 @@ public final class ReadRequest extends SimpleValueRequest<DataReceivedCallback> 
 
 	@NonNull
 	@Override
-	ReadRequest setManager(@NonNull final BleManager manager) {
-		super.setManager(manager);
+	ReadRequest setRequestHandler(@NonNull final RequestHandler requestHandler) {
+		super.setRequestHandler(requestHandler);
+		return this;
+	}
+
+	@NonNull
+	@Override
+	public ReadRequest setHandler(@NonNull final Handler handler) {
+		super.setHandler(handler);
 		return this;
 	}
