--- ble/src/main/java/no/nordicsemi/android/ble/TimeoutHandler.java
@@ -1,13 +1,0 @@
-package no.nordicsemi.android.ble;
-
-import androidx.annotation.NonNull;
-
-abstract class TimeoutHandler {
-
-	/**
-	 * Method called when the request timed out.
-	 *
-	 * @param request the request that timed out.
-	 */
-	abstract void onRequestTimeout(@NonNull final TimeoutableRequest request);
-}
