+++ ble/src/main/java/no/nordicsemi/android/ble/utils/ParserUtils.java
@@ -21,8 +21,29 @@
  */
 package no.nordicsemi.android.ble.utils;
 
+import android.bluetooth.BluetoothDevice;
 import android.bluetooth.BluetoothGattCharacteristic;
 import android.bluetooth.BluetoothGattDescriptor;
+import android.bluetooth.BluetoothProfile;
+
+import androidx.annotation.NonNull;
+import no.nordicsemi.android.ble.PhyRequest;
+import no.nordicsemi.android.ble.annotation.BondState;
+import no.nordicsemi.android.ble.annotation.ConnectionState;
+import no.nordicsemi.android.ble.annotation.PairingVariant;
+import no.nordicsemi.android.ble.annotation.PhyMask;
+import no.nordicsemi.android.ble.annotation.PhyOption;
+import no.nordicsemi.android.ble.annotation.PhyValue;
+import no.nordicsemi.android.ble.annotation.WriteType;
+import no.nordicsemi.android.ble.callback.PhyCallback;
+
+import static no.nordicsemi.android.ble.BleManager.PAIRING_VARIANT_CONSENT;
+import static no.nordicsemi.android.ble.BleManager.PAIRING_VARIANT_DISPLAY_PASSKEY;
+import static no.nordicsemi.android.ble.BleManager.PAIRING_VARIANT_DISPLAY_PIN;
+import static no.nordicsemi.android.ble.BleManager.PAIRING_VARIANT_OOB_CONSENT;
+import static no.nordicsemi.android.ble.BleManager.PAIRING_VARIANT_PASSKEY;
+import static no.nordicsemi.android.ble.BleManager.PAIRING_VARIANT_PASSKEY_CONFIRMATION;
+import static no.nordicsemi.android.ble.BleManager.PAIRING_VARIANT_PIN;
 
 @SuppressWarnings({"WeakerAccess", "unused"})
 public class ParserUtils {
@@ -49,5 +70,134 @@ public class ParserUtils {
 				out[j * 3 + 2] = '-';
 		}
 		return "(0x) " + new String(out);
+	}
+
+	@NonNull
+	public static String pairingVariantToString(@PairingVariant final int variant) {
+		switch (variant) {
+			case PAIRING_VARIANT_PIN:
+				return "PAIRING_VARIANT_PIN";
+			case PAIRING_VARIANT_PASSKEY:
+				return "PAIRING_VARIANT_PASSKEY";
+			case PAIRING_VARIANT_PASSKEY_CONFIRMATION:
+				return "PAIRING_VARIANT_PASSKEY_CONFIRMATION";
+			case PAIRING_VARIANT_CONSENT:
+				return "PAIRING_VARIANT_CONSENT";
+			case PAIRING_VARIANT_DISPLAY_PASSKEY:
+				return "PAIRING_VARIANT_DISPLAY_PASSKEY";
+			case PAIRING_VARIANT_DISPLAY_PIN:
+				return "PAIRING_VARIANT_DISPLAY_PIN";
+			case PAIRING_VARIANT_OOB_CONSENT:
+				return "PAIRING_VARIANT_OOB_CONSENT";
+			default:
+				return "UNKNOWN (" + variant + ")";
+		}
+	}
+
+	@NonNull
+	public static String bondStateToString(@BondState final int state) {
+		switch (state) {
+			case BluetoothDevice.BOND_NONE:
+				return "BOND_NONE";
+			case BluetoothDevice.BOND_BONDING:
+				return "BOND_BONDING";
+			case BluetoothDevice.BOND_BONDED:
+				return "BOND_BONDED";
+			default:
+				return "UNKNOWN (" + state + ")";
+		}
+	}
+
+	@NonNull
+	public static String writeTypeToString(@WriteType final int type) {
+		switch (type) {
+			case BluetoothGattCharacteristic.WRITE_TYPE_DEFAULT:
+				return "WRITE REQUEST";
+			case BluetoothGattCharacteristic.WRITE_TYPE_NO_RESPONSE:
+				return "WRITE COMMAND";
+			case BluetoothGattCharacteristic.WRITE_TYPE_SIGNED:
+				return "WRITE SIGNED";
+			default:
+				return "UNKNOWN (" + type + ")";
+		}
+	}
+
+	/**
+	 * Converts the connection state to String value.
+	 *
+	 * @param state the connection state
+	 * @return state as String
+	 */
+	@NonNull
+	public static String stateToString(@ConnectionState final int state) {
+		switch (state) {
+			case BluetoothProfile.STATE_CONNECTED:
+				return "CONNECTED";
+			case BluetoothProfile.STATE_CONNECTING:
+				return "CONNECTING";
+			case BluetoothProfile.STATE_DISCONNECTING:
+				return "DISCONNECTING";
+			case BluetoothProfile.STATE_DISCONNECTED:
+				return "DISCONNECTED";
+			default:
+				return "UNKNOWN (" + state + ")";
+		}
+	}
+
+	/**
+	 * Converts the PHY number to String value.
+	 *
+	 * @param phy phy value
+	 * @return phy as String
+	 */
+	@NonNull
+	public static String phyToString(@PhyValue final int phy) {
+		switch (phy) {
+			case PhyCallback.PHY_LE_1M:
+				return "LE 1M";
+			case PhyCallback.PHY_LE_2M:
+				return "LE 2M";
+			case PhyCallback.PHY_LE_CODED:
+				return "LE Coded";
+			default:
+				return "UNKNOWN (" + phy + ")";
+		}
+	}
+
+	@NonNull
+	public static String phyMaskToString(@PhyMask final int mask) {
+		switch (mask) {
+			case PhyRequest.PHY_LE_1M_MASK:
+				return "LE 1M";
+			case PhyRequest.PHY_LE_2M_MASK:
+				return "LE 2M";
+			case PhyRequest.PHY_LE_CODED_MASK:
+				return "LE Coded";
+			case PhyRequest.PHY_LE_1M_MASK | PhyRequest.PHY_LE_2M_MASK:
+				return "LE 1M or LE 2M";
+			case PhyRequest.PHY_LE_1M_MASK | PhyRequest.PHY_LE_CODED_MASK:
+				return "LE 1M or LE Coded";
+			case PhyRequest.PHY_LE_2M_MASK | PhyRequest.PHY_LE_CODED_MASK:
+				return "LE 2M or LE Coded";
+			case PhyRequest.PHY_LE_1M_MASK | PhyRequest.PHY_LE_2M_MASK
+					| PhyRequest.PHY_LE_CODED_MASK:
+				return "LE 1M, LE 2M or LE Coded";
+			default:
+				return "UNKNOWN (" + mask + ")";
+		}
+	}
+
+	@NonNull
+	public static String phyCodedOptionToString(@PhyOption final int option) {
+		switch (option) {
+			case PhyRequest.PHY_OPTION_NO_PREFERRED:
+				return "No preferred";
+			case PhyRequest.PHY_OPTION_S2:
+				return "S2";
+			case PhyRequest.PHY_OPTION_S8:
+				return "S8";
+			default:
+				return "UNKNOWN (" + option + ")";
+		}
 	}
 }
